+++ package.json
@@ -80,6 +80,7 @@
     "@types/fs-extra": "9.0.6",
     "@types/fs-readdir-recursive": "1.0.0",
     "@types/gulp-if": "0.0.33",
+    "@types/htmlescape": "^1.1.1",
     "@types/ink-spinner": "3.0.0",
     "@types/jest": "^26.0.20",
     "@types/jsonwebtoken": "8.5.0",
@@ -88,7 +89,6 @@
     "@types/mem-fs-editor": "7.0.0",
     "@types/merge-stream": "1.1.2",
     "@types/mock-fs": "4.13.0",
-    "@types/module-alias": "2.0.0",
     "@types/node": "14.14.22",
     "@types/node-fetch": "2.5.8",
     "@types/parallel-transform": "1.1.0",
