+++ packages/core/src/index.ts
@@ -7,7 +7,7 @@ export * from "./constants"
 export {withBlitzAppRoot} from "./blitz-app-root"
 export {useQuery, usePaginatedQuery, useInfiniteQuery} from "./use-query-hooks"
 export {getQueryKey, invalidateQuery, setQueryData} from "./utils/react-query-utils"
-export {getIsomorphicEnhancedResolver} from "./rpc"
+export {getIsomorphicEnhancedResolver} from "./rpc-client"
 export {useMutation} from "./use-mutation"
 export {invoke} from "./invoke"
 export {getBlitzRuntimeData} from "./blitz-data"
