+++ packages/core/src/server/rpc-server.test.ts
@@ -1,10 +1,11 @@
-import {connectMiddleware, EnhancedResolver} from "@blitzjs/core"
 import delay from "delay"
 import http from "http"
 import {apiResolver} from "next/dist/next-server/server/api-utils"
 import fetch from "node-fetch"
 import listen from "test-listen"
-import {rpcApiHandler} from "../src/rpc"
+import {EnhancedResolver} from "../types"
+import {connectMiddleware} from "./middleware"
+import {rpcApiHandler} from "./rpc-server"
 
 describe("rpcMiddleware", () => {
   describe("HEAD", () => {
