+++ packages/server/package.json
@@ -24,11 +24,7 @@
   ],
   "dependencies": {
     "@blitzjs/file-pipeline": "0.30.6",
-    "b64-lite": "^1.4.0",
-    "chalk": "^4.1.0",
-    "cookie": "^0.4.1",
     "cross-spawn": "7.0.3",
-    "date-fns": "^2.16.1",
     "detect-port": "1.3.0",
     "expand-tilde": "2.0.2",
     "fast-glob": "3.2.5",
@@ -36,13 +32,7 @@
     "from2": "2.3.0",
     "fs-extra": "^9.1.0",
     "gulp-if": "3.0.0",
-    "jsonwebtoken": "^8.5.1",
     "lodash": "^4.0.0",
-    "module-alias": "2.2.2",
-    "nanoid": "^3.1.20",
-    "next": "10.0.7",
-    "null-loader": "4.0.1",
-    "ora": "^5.3.0",
     "parallel-transform": "1.2.0",
     "parse-gitignore": "1.0.1",
     "pirates": "4.0.1",
@@ -51,7 +41,6 @@
     "readable-stream": "3.6.0",
     "resolve-cwd": "3.0.0",
     "slash": "^3.0.0",
-    "superjson": "1.7.2",
     "through2": "4.0.2",
     "vinyl": "2.2.1"
   },
@@ -59,12 +48,13 @@
     "@blitzjs/config": "*",
     "@blitzjs/core": "*",
     "@blitzjs/display": "*",
-    "next-transpile-modules": "6.1.0"
+    "next": "10.0.7"
   },
   "peerDependencies": {
     "@blitzjs/config": "0.30.6",
     "@blitzjs/core": "0.30.6",
-    "@blitzjs/display": "0.30.6"
+    "@blitzjs/display": "0.30.6",
+    "next": "*"
   },
   "gitHead": "d3b9fce0bdd251c2b1890793b0aa1cd77c1c0922"
 }
