+++ packages/server/src/index.ts
@@ -1,4 +1,3 @@
-export {withBlitz} from "./with-blitz"
 export {build} from "./build"
 export {dev} from "./dev"
 export {prod} from "./prod"
@@ -7,7 +6,6 @@ export {normalize} from "./config"
 export type {ServerConfig} from "./config"
 export {resolveBinAsync} from "./resolve-bin-async"
 export {ManifestLoader} from "./stages/manifest"
-export * from "./rpc"
 
 // -----------------
 // For custom server
