+++ packages/server/src/stages/config/index.ts
@@ -40,7 +40,7 @@ export const createStageConfig: Stage = ({config, processNewFile, processNewChil
         cwd: config.src,
         path: resolve(config.src, "next.config.js"),
         contents: Buffer.from(`
-const {withBlitz} = require('@blitzjs/server');
+const {withBlitz} = require('@blitzjs/core/server');
 const config = require('./blitz.config.js');
 module.exports = withBlitz(config);
         `),
@@ -72,7 +72,7 @@ module.exports = withBlitz(config);
       })
 
       file.contents = Buffer.from(`
-const {withBlitz} = require('@blitzjs/server');
+const {withBlitz} = require('@blitzjs/core/server');
 const vercelConfig = require('./next-vercel.config.js');
 const config = require('./blitz.config.js');
 module.exports = withBlitz({...config, ...vercelConfig});
